Added an interpolation class with lerp function.
IGameObjects now require an activate function - IGameObject is becoming more inappropriately named so I may change this in a future revision. This also effects the StateSystem class.

The input class has some extra Update functions that allow the 2d viewing window to be offset from the origin and for the window to change shape.

The mouse has a delta which the distance it's moved in a frame. It also supports the correct position even if the viewing origin and window dimensions change.

Renderer has been given a quick DrawText function. This will draw text without having to do all the setup code, creating a text object and what not. This means it's slower but more convivent to use.

Some more tween functions have been added.